Initial import of helmfile/kind example
No auth is included. I will fix this later.

add helmfile and notes on how to provision a running hello world function

fix a typo and add a few more notes to the README

add another introspective step to the install process

add a note on how to wipe out the kind cluster

add a TODO to add a hapi template

add a note on how to interact with the function

provision the namespaces in code instead of requiring an extra step

remove need to manually create namespaces

fix erroneous inclusion of manually creating namespaces